# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lotus Funds API
  version: 1.0.0
  description: |
    Documentation for the existing Lotus Funds Express API.
    Response envelopes intentionally differ where the implementation differs.
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/performance:
    get:
      tags: [Performance]
      summary: Calculate monthly performance for the authenticated RA
      security:
        - bearerAuth: []
      parameters:
        - name: month
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
            example: "2026-07"
      responses:
        "200":
          description: Performance metrics envelope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
  /notifications:
    get:
      tags: [Notifications, Admin]
      summary: List Admin notifications
      description: Allowed roles are ADMIN and SUPERADMIN.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification data returned by the controller
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/subscription-plans:
    get:
      tags: [Subscriptions]
      summary: List public subscription plans
      security: []
      parameters:
        - name: audienceType
          in: query
          schema:
            type: string
            enum: [RA, BROKER]
      responses:
        "200":
          description: Subscription plans
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/ValidationError"
...
  /api/audit-logs:
    get:
      tags: [Audit]
      summary: Query audit logs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Audit log query result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Search:
      name: search
      in: query
      schema:
        type: string
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated user does not have the required role or subscription access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        error: { type: string }
        field: { type: string }
        nextStep: { type: string }
      additionalProperties: true
...
    PerformanceResponse:
      type: object
      required: [success, month, metrics]
      properties:
        success: { type: boolean }
        month: { type: string, example: "2026-07" }
        metrics:
          type: object
          properties:
            total: { type: number }
            active: { type: number }
            exited: { type: number }
            accuracy: { type: number }
            profit: { type: number }
            adverse: { type: number }
            strike: { type: number }
            sl: { type: number }
            early: { type: number }
            rr:
              type: number
              nullable: true
            last:
              type: array
              items:
                type: string
                enum: [g, r, n]
          additionalProperties: true